refactor(caching): Redo search caching to pull out variables that are un-pickleable
It turns out that the facet_fields variable cannot be cached anymore, so I redid the way we do
caching by moving the cache closer to the queries. This, unfortunately, breaks up what was
previously a larger object in the cache into as many as three smaller ones. One for each solr query
on the home page (oral args and opinions), and one for stats. That's not so bad; just more RTs to
Redis to load the page.

What's more annoying is the damage this does to the already complex code. Now, instead of (just)
doing the queries in the do_search() method, we also have a layer of caching on top that can turned
on or off as the caller pleases. Logically, that's...OK, but the complexity and ugliness it adds to
the code is real. We now have a 120 line function here.