fix(search caching) Another attempt at this
This bug only happens once every six hours when the cache expires. Makes it particularly fun to deal
with. In any case, I did some testing on the server and it looks like we can get away with cached
search results *only*, so hopefully this will accomplish that. Each fix to this bug has gotten
successively narrower in what we cache and at this point it's about as narrow as possible, so this
should do it.