fix(celery): Upgrade Celery to 4.3
This is annoying, as described in my comment here:

https://github.com/celery/celery/issues/4878#issuecomment-491528733

Basically, Celery 4.2 and 4.3 (the most recent) have bug:

https://github.com/celery/celery/issues/4878

Which makes it not work with CONN_MAX_AGE>0 in Django. And I just discovered that Celery 4.1 has
bug:

https://github.com/celery/celery/issues/3808

Which makes it overload Redis until it crashes. Previously I fixed issue 4878 by installing 4.1, but
now that we know about issue 3808, we have no choice but to go back to 4.3, and to disable
CONN_MAX_AGE. This commit does that.