fix(donate): Align qr code with address
We have two addresses in Coinbase that work. No idea why they're both listed here, but this aligns
them properly in response to an email where somebody thought we might have been hacked to have the
wrong image or text or something. We weren't, obviously, but this at least ensures that nobody is
confused again.