fix(donate): Update template attributes to set active radios properly
This broke in Django 1.11, with a somewhat poorly documented deprecation. Django took away a class,
and replaced it with something else, so an attribute of that class that we relied on
("choice_value") was killed while only mentioning that tha class itself was gone.

The issue manifested as all of the radio buttons being active in the front end on the donate page.
This fixes them so they only show as active if they're actually supposed to.