fix(apache): Pull out bare domain redirect to www
With the addition of ssl-matomo.courtlistener.com.conf, we had two ServerName directives that
responded to requests on port 443:

 - www.courtlistener.com
 - matomo.courtlistener.com

This caused the issue that when somebody came on the bare domain — courtlistener.com — Apache would
have to make a decision about which VirtualHost for port 443 would handle the request. Since neither
VirtualHost had a matching ServerName, Apache just chose the one that came first alphabetically,
matomo. This fixes that issue by creating an explicit virtual host for the bare domain on SSL,
similar to the one we have for port port 80. So now we have the following matrix:

 - matomo.courtlistener.com:443 --> Proxies to matomo server
 - www.courtlistener.com:443    --> Serves requests
 - courtlistener.com:443        --> Redirects to www (new)
 - matomo.courtlistener.com:80  --> No config, but your browser should redirect you due to HSTS
 - www.courtlistener.com:80     --> Redirects to https.
 - courtlistener.com:80         --> Redirects to https://, which redirects to https://www

Probably a better way to handle this, and I don't love that we have three files repeating the SSL
config, but doing it via an include is probably unwise with Let's Encrypt doing weird stuff to files
on a semi-regular basis. So...here we are. It works and for now it's not that complex.