fix(alerts): Suppress an irritating warning
This has been a warning in the system for years. It happens whenever we send a big query to Solr:

https://github.com/lugensa/scorched/blob/master/scorched/connection.py#L212-L214

The reason you'd put this is to make sure that the caller knows that their query is so big that it
cannot be cached in Jetty/Tomcat, which is fine, and so the big solution is to make a longer query
be allowed by the servlet, but the cache isn't that important here, so I'm not sweating it.