fix(free opinions): Handle failures more aggressively
Previously, if we got a request failure, we aborted, and if we got anything else, we crashed. This
makes it so either of these errors are essentially logged, but ignored. Since we can't easily fix
the 6.3 bug, this at least lets things finish. Otherwise, it'd crash and we'd never get anything,
which is really bad.