fix(admin): Ensure that when we seal items, Solr is updated
I've no reason to think this is happening, but we currently have a bug that even after a sealed item
has been removed from CL, the contents of that document are still available in Solr. This bug
ensures that we update Solr after we update items via the admin pages, thus fixing this issue.

Fixes: 981