fix(apache): Allow munin to access apache again
This got broken in a0314c38faa0bea47926c17f3144992554ed45ac, which introduced the matomo subdomain.
When we did that, because of virtual hosts, all requests for /server-status started going to the
matomo subdomain (because it comes alphabetically before www). So, to fix this, we need to allow
access to this Location from any internal IP address, and we need to update the munin configuration
to access courtlistener.com instead of localhost.