fix(apache): Allow munin to access apache again
This got broken in 78a91a2bf01e1621666fb0d36e567bb46fd3a510, which introduced the matomo subdomain.
When we did that, because of virtual hosts, all requests for /server-status started going to the
matomo subdomain (because it comes alphabetically before www). So, to fix this, we need to allow
access to this Location from any internal IP address, and we need to update the munin configuration
to access courtlistener.com instead of localhost.