refactor(tests): Use standard attribute, "live_server_url" for integration tests; tweaks compose
This adds selenium to the compose file and switches the container names from underscores to dashes.
The latter is important because this is not a valid domain: cl_django. This is: cl-django. And
django complains if you domains are overly wrong.

This also adds several deps to our django image. Unfortunately these are needed to Celery can run
synchronously during tests. If you try to run Celery async during tests, it'll connect to your non-
testing DB, and any effort to make it not do so will make you crazy.