fix(tasks): Add retry mechanism for failing links
This seems to fail a lot in nysb, but I'm not sure why. They *would* have some really big
bankruptcies there, which *could* have big claims tables, which *might* cause the DB to timeout, but
when I just looked at a failing one, the claims table had — drumroll — one entry.

So, not sure what's up with this, but I'm going to throw some retries at it and hope that fixes it.