feat(client work): Refactor export control script and add support for stage 2
This pulls out the task handling code and generalizes it so that a second task can be applied. Our
original task of parsing the file provided by FDD should still work, but now we should also be able
to support any type of CSV, provided it has the needed data in it somewhere.