fix(docker): Switch to bridge networking
Bridge networking can't be used in a swarm because it cannot span multiple machines, but I'm
switching to it because I can't get overlay to work on my new laptop. Fortunately, we're not
spanning multiple machines, so this shouldn't have any major consequences.

The error I've been getting, and my attempted solutions to it are here:

https://stackoverflow.com/questions/57701373/cannot-use-docker-compose-with-overlay-network