fix(docker): Update compose to use customizable users
This is necessary because otherwise getting permissions right on the volumes becomes *really*
difficult. So instead of attempting to align the user IDs inside the docker containers with the
permissions on all our files, we have an escape valve: Just set the user you want the docker
container to run as via an env var.

This fixes the various complexities solved, sort of by blog posts like this one:

https://medium.com/@nielssj/docker-volumes-and-file-system-permissions-772c1aee23ca