fix(docker): Make the makefile build the correct image
It was, unfortunately, and very frustratingly building the task image, which is very similar to the
django image (too similar really), and when it did that, things just wouldn't start correctly. This
shouldn't be a surprise because if you build using the wrong Dockerfile and then push that to
hub.docker.com, then pull it and start using it like it's something else...well, it's not going to
work.

So, this little fix and little error were the cause of a lot of headaches, but happily they're
fixed.

For now.