refactor(crypto): Cleanup and combine all sha1 and sha256 code
This was a big mess, as you can see by the size of the commit. One nice
benefit though is that we were dumbly generating SHA256 hashes from
audio files previously. The way we'd do it pulled the whole file into
memory. Why? Who knows. But now we just do a UUID4. So much better!