fix(payments): Add XERO integration to the stripe callback function
A few weeks ago, I started letting people pay their invoices from XERO.
As people have begun doing this (great), we have begun getting callbacks
from stripe to this webhook endpoint. Alas, when that happened, we did
not have any information in the CL DB about the payment. This makes sure
that we add that info before trying to process the stripe callback.