fix(recap alerts): Disable retry for RSS changes
This task looks for changes to the RSS feeds, and currently has a retry
of five times. Because it also has a lengthy timeout, it can really tie
up a celery workers when a site goes down. What's worse is that trying
five times, each with a five minute wait means that there's a good
chance you'll still be trying to retry the *last* task when the next
scrape of the site comes along (I think). So...if a site goes down, or
as in the case today, multiple ones do, this can totally tie up Celery.