fix(req.txt): Update kombu and django-extensions
Django Extensions was failing on our version of Django and had to be
updated.

Kombu is a different, darker story. The reason it needs to be changed is
because we didn't previously have it in place, and instead relied on
celery to get us the right version. However, for no apparent reason,
celery recently started requiring a more recent version of kombu. That
version, it turns out crashed, with this bug:

https://github.com/celery/kombu/issues/1063

So...we had to specify which version of kombu we wanted, and that's what
we do here. Ugh.