feat(recap fetch): Launch RECAP fetch
There is some complicated code here, so I better explain what's going
on. The big picture is that we need to take a request from the user and
route it through a chain of celery tasks. Those tasks, however, vary
based on the request the user made. Obviously, if they request a PDF or
a docket, that makes a big difference, but it's perhaps less obvious
that some requests can trigger additional PACER look ups that others
wouldn't. So the trick here is to compose the chain with all the right
parameters and tasks according to the request the user made.

Let's talk code paths, because that's where this gets complex. There are
two major paths: Docket and PDF. PDFs are obvious, but dockets are not.
For the dockets, we can get requests in three formats:
 - docket number + court
 - pacer_case_id + court
 - CL docket_id
The tasks we take vary depending on whether we already have the
pacer_case_id, or whether the user provided it to us. If we already have
it, we just go and get the docket. If we don't, we have to get it first,
then go get the docket. Once we've gotten the docket, we have to save it
to Solr, then mark the task as complete.

What's missing here? The big thing that's going to be an issue is
reporting back to the user when things go wrong. Right now we suck at it
because we're using all the old tasks that were designed just to log
things. Now we need to be putting errors into the db instead, but we're
not. I suspect a better architecture would merge all of our task
processing into one big thing. So, for example, next time we want to do
a customer job, instead of calling the tasks directly, we'd create tasks
using this API, and it'd do the job for us. Hm.