fix(stats): Allow "AnonymousUser" as a user in redis
Earlier this year, I made a change to upgrade the version of the redis
library that we use, it no longer allowed None values and instead required
"None" values. That was fine, but along the way I changed the way we
collect stats in redis to put "AnonymousUser" as a key instead of "None".

So...this fixes our stats compiler to understand either value.