feat(black): Add a pyproject.toml file with black rules
This format is messy, but `include` is a regular expression that matches
py and pyi files in `cl` OR `OCR`. `exclude` excepts migration files.

Note that it's important to only include *\.pyi?$ files with this b/c otherwise
black can and will reformat csv files and json files that are valid Python!