feat(black): Reformat entire codebase using Black
This is the biggest commit in the history of CourtListener. It reformats
the entire codebase using the black code formatter. Black is pretty darned
good, and with the default `--safe` mode, I believe it ensures that every
resulting file still compiles, so this should be safe.

However, I did make a few changes:

1. When Black encounters code like this:

        log("My long thing "
            "goes on for ages.")

    It converts it to:

        log("My long thing " "goes on for ages.")

    That's not great, so I fixed all of those manually.

2. I noticed two places where the above limitation in black actually
   identified bugs inside long lists. lasc/admin.py is one, and one
   of the court parsers was the other. In both cases, they were missing
   a comma in a way that caused a little damage, so I just fixed them.

3. I read through a lot of this diff to satisfy myself that it was fine. As
   I did so, I identified a few places where the formatting from Black wasn't
   desired. court_regexes.py is the most obvious place. In those instances, I
   added the # fmt: on/off switches and reverted Black's changes.

With this commit, and the associated Github action that enforces it, we're
"committing" (ha, sigh) to having well-formatted code, or at least code that
satisfies the machine. From here on out, assuming CI works for this commit, if
Black fails, the code has failed and will not be merged.

Whew!