fix(payments): Don't create a donation record if we already have one
Honestly, I don't know how we're winding up running code twice. I think what
is happening is that Stripe fails for an unrelated reason, and then does a
retry. Each time it retries it's able to make more Donation objects, and it
just makes more and more of them. I think by fixing this we should be able to
find the other error downstream in the callback handling from this point.