feat(migrations): Add depth field to citations model
The migration here a bit complex. It turns out that adding this field does not one,
but two bad things in Postgres:

1. It uses a default field, which means the table has to be rewritten.

2. It sets an index, but not concurrently, which causes a table lock until the index
   is built.

Note also that we do this via two SQL commands and a transaction with two more. The
reason it's broken up this way is because data and model changes can't be in the same
transaction. This way you have a model change, a data change, and then a model change.