feat(js): Upgrade jquery to 3.x from 1.11.x
Yes, I know jQuery is old, but this version of it is needed for the latest
version of bootstrap, which I'll be landing shortly. The upgrade here is fairly
straightforward:

1. Upgrade .click() and similar to .on("click",

2. Remove the fallback in base.html that only makes sense when using a CDN, but
   which we've had for ages.

The process for this was to use the jquery-migrate.js file. With it in place in
base.html, I browsed most of the site. As I did, it'd throw errors and I'd go
grep them down. I browsed around all the parts of the site that I could think of
that have dynamic content. There's not a super obvious way to do this, but I was
pretty thorough.