fix(find_citations): Typical, Atypical tax court citations failing
We built a method to handle atypical tax court cases.

For example: T.C. Memo 2019-12.  This gets saved and stored as
2019 T.C. Memo 12 in our database.  But running 2019 T.C. Memo 12
into our system recognizes 2019 T.C. Memo 12 as a atypical tax case
even though it is now in a typical format.  Processing a typical
format in the atypical (specialized) way caused the crash.

A test was added for the new format we now make