fix(search): Use the entire dict returned by do_search
This reverts an earlier change that made our render_dict.update() call more
elegant, but in so doing, it removed all the variables returned by do_search
except for the `results` variable. That actually didn't break much! But it did
make it so that the number of jurisdictions didn't show up on the homepage.

Not great. Fixed.