fix(css): Eliminate tiny line between bell and main input box
This was a bit of a painful one. Prior to this change, when you had got the bell
icon next to the input box, you'd see a 1px wide blurred vertical line between
the two elements. This was created by the box shadow, and setting the size of the
shadow to 0px didn't fix it because 0px gets blurred too! The solution, then is
to shift the direction of the input box's blur to the left and to shift the
shadow of the bell icon to the right. That way, you get a nice shadow and you
don't wind up with a blurred 0px line in the middle.

(This also removes some vendor prefixes and moves a couple things around.)