fix(tests): Remove files in assets/mp3
These files cause lots of fun issues when they're deployed because they
overwrite the mp3 directory on the live server. That breaks links to mp3 on the
front end in really bad ways. This issue first came up as a broken test as
indicated in a6c32388d4431e196a3bc533f9438fb6fe1b9a31. The fix there was to add
the ander_v._leo.mp3 fixture, which hadn't been necessary before, but by adding
that file, we created the aforementioned breakage.

So this commit removes that file again and updates the fixture to put it
somewhere else. I still don't know why things broke prior to
a6c32388d4431e196a3bc533f9438fb6fe1b9a31, but it's interesting to look at how
they did. We currently have two CI's set up per commit, and what's really weird
is that in 192450e03988516c8ec27526fd92cbc15ebbdab6, only *one* of the two
identical builds failed. That hints at a timing error or something of *that*
kind, but I couldn't find anything like that in our tests, so I'm baffled.

After that commit, the test started failing consistently, which is why I
put in a fix.

This is a weird one. Hopefully this is the fix we seek.