fix(tests): Remove files in assets/mp3
These files cause lots of fun issues when they're deployed because they
overwrite the mp3 directory on the live server. That breaks links to mp3 on the
front end in really bad ways. This issue first came up as a broken test as
indicated in 561fdfb764ecadc6c49006c95671e6bc2badf806. The fix there was to add
the ander_v._leo.mp3 fixture, which hadn't been necessary before, but by adding
that file, we created the aforementioned breakage.

So this commit removes that file again and updates the fixture to put it
somewhere else. I still don't know why things broke prior to
561fdfb764ecadc6c49006c95671e6bc2badf806, but it's interesting to look at how
they did. We currently have two CI's set up per commit, and what's really weird
is that in 5b933106dd7946f5124870e9ad4887afd8bebf8a, only *one* of the two
identical builds failed. That hints at a timing error or something of *that*
kind, but I couldn't find anything like that in our tests, so I'm baffled.

After that commit, the test started failing consistently, which is why I
put in a fix.

This is a weird one. Hopefully this is the fix we seek.