feat(css): Reduce the padding between the alert bell and the query box
This should make it possible to have longer queries, particularly on narrow
viewports. Further, it should make it so clicking to the left of the bell in the
query box shows the blinking [ I ] cursor and allows you to work with the box.
Previously clicking in that region did nothing.