fix(recap): Create a new PROCESSING_STATUS value for when we need more info
I wasn't happy with the way the attachment page failed prior to this. My options
were to either say that it failed due to INVALID_CONTENT or just to say it
FAILED. Neither really captured this. INVALID_CONTENT is for when somebody's API
request was wrong. FAILED is for when something goes wrong.

In this case, things worked properly — they just needed more info to actually
work.