feat(recap): Add better error messages to PacerFetchQueue items
This was a bit tricky, but the heart of it is that the fetching task now
hands its status back up the chain and there's a new task that just handles
the response from the fetching task.

This isn't *wonderful* per se, but it works and the fetching task is more or
less untweaked, which is good.