feat(recap): Make a single task that can handle the whole attment fetch
This is both a simplification and a complication. Simplification because it
uses code from elsewhere. Complication because it adds more robust status
handling to address #1248.

Happily, it no longer creates ProcessingQueue objects, and it no longer needs
a celery chain to run. Sadly, the new task for this is a bit of a doozy.