fix(tests): Clear the cache after each API test
This fixes leakage across tests. If you're not careful, you can cache
the results of a query in one test, and then your query count test
will be off because it loads from cache. Cross-test pollution is fun
for all ages.

I tried a few other approaches before landing on this one:
 - Mock the cache_page function. Strangely this didn't seem to work.
 - Use @override_settings to swap in either the DummyCache or the
   LocMemoryCache. This didn't work either because at first
   authentication became an issue (it wouldn't work with DummyCache),
   then the cross-test pollution resumed, because things would survive
   in LocMemoryCache.

So...I thought of a different thing. Nuke the cache!