refactor(js): Factor out the alert creation JS
One little trick here. We still need to set the variable in the HTML
template, since the value comes from Django. As a result, we just set
a global variable that the JS can grab.

The other way to approach this is to call the function itsel from the
HTML template and to pass the variable that way. That's better from a
"Globals-Bad" perspective, but it seemed harder and more verbose, so I
opted not to do that.