feat(alerts): New docket alert page
This lands the view, URL, and template for creating docket alerts from
RECAP. The idea is that users will get a button on every docket they
see in PACER. Clicking the button takes them to this page, which gives
them the ability to *very* easily create alerts.

There's not much that's new or clever here. Mostly this just takes
bits and pieces of code from the docket page and elsewhere and puts
them all together into a new page. I do like the way it came out
though.