fix(docker): Simplify DB connection info
This should be a good fix for those on OSX, but it requires that those
of us that relied on the CL_POSTGRES_RUN_DIR variable update our code.

The update should be simple though. Just add the HOST parameter that's
now in 05-private.example to your private settings file.