feat(api): New views for JSONVersions and Visualizations
Mostly normal stuff. The one trick that was needed was to override the
create function. This seems like overkill, but the issue is that if
you have generating code in your perform_create that can fail,
there's no particularly good way to throw an error from that code. In
this instance we build the network in our perform_create function.
Doing that can throw a couple different errors, namely, too_many_nodes
or too_few_nodes.

To know that those errors happened we override create, which can
actually return appropriate responses according to the error messages.