fix(search): Add docket entry metadata when indexing documents
This is kind of a disappointing bug. I obviously meant to add the
docket entry data to item when adding it to Solr, but then made a
typo, and instead added the recap document data twice. Shoot. The good
news is that we rarely index by docket, so that's why this bug has
been around so long.