feat(mergers): Add support for datetime date_filed fields
The long term goal is to support tz-aware datetimes in the DB. The
first step towards doing that is to start collecting those in
Juriscraper. A version I plan to land later today will do that for
RSS feeds, so this commit makes sure that CL won't break when that
happens.

Down the road, we need to start keeping datetimes in the DB instead
of dates. When we get there, we should just be able to save the
datetime objects directly without messing around. I think!