fix(scrapers): Handle it if the a free opinion doesn't exist
I'm not sure why this happens, to be honest. We have a simple loop
that creates tasks to go and download free opinions from PACER. But
somehow the items to download don't exist when the task is run, even
though they did when the task was created. As far as I can tell, these
items aren't deleted anywhere else, but clearly they get deleted
somehow.

Anyhow, the fix is simple. If the row doesn't exist, we just don't
bother with it anymore.

Fixes COURTLISTENER-1E