feat(iquery): Allow back-crawling of iquery
In COURTLISTENER-1Q, I identified that judge names were getting
screwed up by the rsplit command being improperly used. Juriscraper
has the fix for that, and we need a way of going back and fixing the
data that we crawled with that bad information in the meantime. This
new feature supports that by allowing us to query Matomo for
additional days previous to today.