fix(caching): Make homepage survive a broken cache
If it happens that we cache a visualization for the homepage, but then
that viz is deleted, we'll have it in cache, but we'll lack the JSON
to back it up. That's bad and could cause the homepage to crash at
this very moment! Luckily our tests caught this (sort of). The fix is
to cache the error in the model and to display things well on the
homepage when it happens.

Fixes #1310