fox(alerts): Fix timezone errors in alert tests
Tricky one, but the length docstrings should explain it. Basically,
django does a fair bit of date conversions when you put datetimes in
the DB, and I was calling date() a lot, which also does tz
conversions, and it was a bit of a mess.

The new path forward is that we compare dates, after they're
normalized to the server's timezone. This should prevent the tests
from failing at certain times of day. I tested this locally while
tinkering with my laptop's time, so I think it's in good shape.