refactor(fetch): New fetch method to handle errors much better
The problem with the old chained approach was that errors weren't easy
to catch. You could try, sure, but if the error happened outside the
main function, you'd struggle to catch it.

This introduces a large new function which should have the same
functionality while being much more capable of catching errors
properly for our users.