feat(pacer): (De)Compress RSS before saving it or reading it
I looked at a few different ways to accomplish this, but decided to go
with the simplest. The other approachs were to use a special file
field package that had compression utilities, or to make a custom
storage handler based on the default one that did (de)compression on
the fly.

Adding another package for this felt dumb. I tried doing the custom
storage class, but it was a fair bit of finicky work that seemed
fragile. For example, you have to hvae an `_open` method that returns
a file handle, but bz2 doesn't think that way. Should we instead
create a temp file? Eh, who knows.

Other considerations:

1. We talked on slack about different compression algorithems. BZ2 is
the second slowest, but it does reward with very good compression. It
beat out gzip and zip and taring them gzipping. I looked a bit at
Brotli, but it was hard to install except through pip and the Python
bindings were undocumented. Plus, bz2 doesn't leave it a lot of room
to improve on.

2. Compressing file systems like ZFS or BRTRFS might help here too.