fix(pacer): Always save last build date
This is a small tweak that creates a big fix. In #1311 we discovered
that a LOT of parsing was being done for RSS feeds, and a LOT of
redundant data was being saved to disk that shouldn't have been. The
reason was because we didn't remember to save the last build date to
the DB when the feed was unchanged.

That resulted in the following pseudo-code:

1. Check the feed. Has the build date changed vs the latest one in
   the DB? Yes. Save the latest bulid date to the DB and download and
   parse the feed.

2. Wait awhile...check the feed. Has the build date changed since the
   last time? No; move on. Save to the DB that we checked the item
   and it was unchanged but omit the build date column.

3. Wait awhile...check the feed. Has the build date changed since hte
   last time? Yes! The last time has a value of None for the build
   date and now we have a different build date. Download and parse it;
   save the build date to the DB.

4. Repeat step 2, then 3, then 2, then 3....

So every other time we checked it, we'd find that the build date had
changed. Lovely!

Anyway, the fix is simple, as you can see. Just make sure that if we
are able to get a build date, we save that value to the DB for future
iterations.