fix(cache): Let the caches breathe a bit
I'm really bothered that I missed this until now. Django's default
cache size is 300 items, which is, frankly, almost nothing for us.
This makes our caches much bigger. With Redis, this is going to
take some memory, but it shouldn't be much, really. With the DB,
it'll take some space on disk, but 250,000 indexed items in a DB is
nothing.

One area this should fix is sitemaps. I thought we were caching them,
and technically we were, but since our cache was so small, they were
just getting evicted all the time.

Ugh!