fix(tags): Move from content types to m2m
Lots of warning everywhere about using generic foreign keys. I went
for it anyway, because tags are exactly the kind of thing they're for,
but in the end, they're a pain. Plus, if you want to do a m2m (as we
do here), they seem to get even more painful — or maybe they're not
even possible, I didn't bother to find out.

Partly, I also wanted control over the through table, so I could
create an API for it. That's needed so you can actually tag things.