fix(pdfs): Don't wait for Matomo to respond when serving PDFs
This modifies our decorator to make it work either with or without
parameters. If you supply parameters, they are used, but not it works
too.

More importantly, this adjusts our PDF handling to be faster. Matomo
now has 10ms to get a response to us, which is *fine*. If we don't get
a response, we just serve the file, but assuming matomo can keep up,
we'll still track it.

The biggest thing I don't like about this approach is that if Matomo
starts throwing lots of errors, we might miss that, but I think that's
an OK compromise.