fix(settings): Attempt to disable admin emails again
The problem I'm currently having is that Sentry isn't sending alerts
when warnings are raised. My hope is that it's because our logging
config is suppressing the emails.

The previous commit broadens the logging to log more stuff, and this
one, hopefully, prevents default Django error emails from happening,
leaving only the Sentry ones.