fix(meta): Add date_modified=now() to relevant update calls
This is the kind of Django wart that takes a long time to notice. I
only caught it today because I noticed that when I updated innumerable
FJC IDB rows, the date_modified stamp that's exposed in the front end
wasn't updating.

When I tried to sort out the cause, it turned out it was because I was
using the `update()` method, which doesn't update `auto_now` fields.

This sort of makes sense — it just does an SQL UPDATE, and that's
that, but it breaks with a lot of how Django just does the right
thing.

This goes on the long list of things to look out for. Bleh.