feat(tags): Provide link to tag page on each tag
Two pieces here:

1. Add the link itself as a span. Use a span because we're already
inside a link, so we can't use a link like we would normally.

2. Pass the user information down from the DOM into the ListItem, so
that we can use it to make a link. I don't love the tight tethering
here. For example, if we change the link to remove the user name, we
have to adjust props being passed and whatnot. Meh.