feat(tags): Don't show public tags to other users
By default, the API returns public tags and your tags. So if we don't
filter them, you wind up seeing other people's public tags in your UI.
That's confusing, so don't do that.

This can be reproduced by creating two users. With one of them, create
a public tag. With the other, check for that tag in the UI. You
shouldn't see it.