fix(451): Block sealed case from search engines
Use Apache so that it can be unblocked when we convince the judge this
was crazy. If, in many years, you are looking at this line and
wondering if it's needed, that means we lost our fight with the judge,
and the case stayed sealed. If that happens, you can remove this line,
but only after you actually remove the case from the DB.