fix(api): Properly hide uploader field
The idea here is that you don't want the uploader of a RECAP document
to be known. Everything from the extension will be uploaded by the
"recap" user, but sometimes we have other uploaders and their privacy
should be maintained.

In older DRF, you did that by excluding the field and by marking it as
hidden. These days, hidden appears to be enough. Verified by checking
the API results to make sure that the field doesn't appear.