fix(drf): Upgrade DRF filters
1. DjangoFilterBackend --> RestFrameworkFilterBackend (just to screw
with us).

2. Dependencies get upgraded. Note that for the moment, drf-filters
requires a dev version to work with django-filter > 2.0.
django-filter > 2.0 is required to work with a version of DRF that
works on py3, so here we are. (django-filter > 2.0 came out more
than two years ago.) The dev version of drf-filters appears to work
fine. In their milestone for proper release of it, the only remaining
work is on documentation.

3. filter_class --> filterset_class (I guess this is more consistent
with...something)

4. Tests have to get upgraded. Two changes here. First, for fields
that require choices, invalid options return a 400 status code instead
of zero results. This is generally good, but requires that our tests
be upgraded. Second, if you use two related filters from the same
related object, they are AND'ed instead of OR'ed. This makes sense to
me, and is how they'd work on the endpoint itself, but it required a
tweak to a test.

5. More tests for swagger and the OPTIONS http requests.

6. One more HiddenField that needed to not be excluded (in
PacerFetchQueueSerializer)