fix(cleanup): Keep things in unicode
I expect there's a reason this was so funky, but I noticed it during
the pull and decided to fix it. The approach here includes a test and
keeps things in unicode instead of converting them back and forth into
bytes. We shouldn't need to do that much, if at all, anymore.