fix(apache): Set our locale for mod_wsgi
This was a truly annoying issue. The heart of it is that mod_wsgi
doesn't use utf-8 by default and instead uses ascii. You can change
this using the lang and locale configurations of the WSGIDaemonProcess
directive.

Normally that's easy, but because we're using a slim docker image, we
don't have the locales that are normally installed, so we have to
install them in docker before we can use them in Apache.

This began as a UnicodeDecodeError when importing a json file from
reporters-db. Two further references with details are:

https://github.com/GrahamDumpleton/mod_wsgi/issues/625
https://stackoverflow.com/q/64564924/64911