feat(urls): Simplify binary file handling
This accomplishes a few things. It:

1. Swaps over to nginx's xsendfile tech, which they call X-Accel.

2. Stops serving a bunch of file types from the application layer.
From now on, these file types (doc, docx, wpd, txt, html, and mp3)
will get served directly from nginx. This means we lose their
tracking, but it will make them much faster and reliable. This also
means that we can't set per-request search engine blocking on these
file types, so we just block them entirely. That's the more private
thing to do, and it might even help our SEO not to have a PDF and an
HTML version of every file.