fix(docker): Add bulk data dir to gunicorn
The X-Accel header combined with gunicorn as a proxy is neat because
it means that to serve a PDF, even with special heads and whatnot, the
application server (gunicorn) doesn't need access to PDFs. They're
eventually served by nginx, so they're not needed by gunicorn. Neat!

I thought that might mean that we could get away with not giving
gunicorn access to the PDFs. Well, that's almost true, but crawlers
want thumbnails, they get made by the application server the first
time that crawler tries to access the page. Alas, the application
server needs access to the PDFs to do that. Oh well.

This provides that access.