feat(docker): Add selenium to nginx server compose
Selenium is needed for various scraping cronjobs. I wanted it to live
elsewhere and be independent from the docker compose file, but
creating and tweaking a second docker compose, or somehow making it
work without one was seeming pretty daunting. To the contrary, this is
quite easy.

Juriscraper 2.1.0 supports remote webdrivers, so this provides one it
can connect to (hopefully!).