fix(logging): Switch to console logging by default
I am getting a strange error on my dev machine that:

--- Logging error ---
Traceback (most recent call last):
  File "/usr/local/lib/python3.8/logging/handlers.py", line 69, in emit
    if self.shouldRollover(record):
  File "/usr/local/lib/python3.8/logging/handlers.py", line 186, in shouldRollover
    self.stream.seek(0, 2)  #due to non-posix-compliant Windows feature
io.UnsupportedOperation: underlying stream is not seekable
Call stack:
  File "/opt/courtlistener/manage.py", line 15, in <module>
    main()
  File "/opt/courtlistener/manage.py", line 11, in main
    execute_from_command_line(sys.argv)
  File "/usr/local/lib/python3.8/site-packages/django/core/management/__init__.py", line 364, in execute_from_command_line
    utility.execute()
  File "/usr/local/lib/python3.8/site-packages/django/core/management/__init__.py", line 356, in execute
    self.fetch_command(subcommand).run_from_argv(self.argv)
  File "/usr/local/lib/python3.8/site-packages/django/core/management/base.py", line 283, in run_from_argv
    self.execute(*args, **cmd_options)
  File "/usr/local/lib/python3.8/site-packages/django/core/management/base.py", line 330, in execute
    output = self.handle(*args, **options)
  File "/opt/courtlistener/cl/alerts/management/commands/cl_send_alerts.py", line 112, in handle
    self.send_emails(options["rate"])
  File "/opt/courtlistener/cl/alerts/management/commands/cl_send_alerts.py", line 229, in send_emails
    logger.info("Sent %s %s email alerts." % (alerts_sent_count, rate))

This appears to be because I used a symbolic link in our DockerFile to
map the log file to /dev/stdout (which isn't seekable). The general
goal with docker is to send your logs to stdout, which the above
approach (borrowed from nginx, I might add!) does! Alas, it doesn't
work here, so instead we switch to the logging StreamHandler, which
does work nicely.

This also includes a tweak to ensure that on dev machines we have
verbose logs and on prod we do not.